orchid-project/backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
google-generativeai
beautifulsoup4 